requirements.txt
fastapi
uvicorn[standard]
pydantic
uiautomator2
pillow

